moved original code to "old" dir, fixed errors so works with cli
