+++ old/package.json
@@ -1,0 +1,83 @@
+{
+  "name": "conf-twitter-bot-server",
+  "version": "0.0.1",
+  "description": "CS Conference Twitter Bot Server",
+  "author": {
+    "name": "Stefan Marr",
+    "email": "git@stefan-marr.de"
+  },
+  "license": "MIT",
+  "dependencies": {
+    "@koa/cors": "^3.3.0",
+    "@koa/router": "^11.0.1",
+    "@types/html-minifier-terser": "^6.1.0",
+    "@types/user-agents": "^1.0.2",
+    "axios": "^0.27.2",
+    "dotenv": "^16.0.1",
+    "html-minifier-terser": "^7.0.0-beta.0",
+    "http-status": "^1.5.2",
+    "jsdom": "^20.0.0",
+    "koa": "2.13.4",
+    "koa-body": "4.2.0",
+    "koa-logger": "^3.2.1",
+    "koa-mount": "^4.0.0",
+    "koa-session": "6.2.0",
+    "koa-static": "^5.0.0",
+    "mustache": "4.2.0",
+    "node-cron": "^3.0.1",
+    "node-fetch-cookies": "^2.0.3",
+    "twitter-api-v2": "1.8.0",
+    "user-agents": "^1.0.1100"
+  },
+  "devDependencies": {
+    "@types/datatables.net": "1.10.21",
+    "@types/jest": "27.0.3",
+    "@types/jquery": "3.5.11",
+    "@types/jqueryui": "1.12.16",
+    "@types/jsdom": "16.2.13",
+    "@types/koa": "2.13.4",
+    "@types/koa__router": "^8.0.11",
+    "@types/koa-bodyparser": "^4.3.7",
+    "@types/koa-session": "^5.10.4",
+    "@types/mustache": "4.1.2",
+    "@typescript-eslint/eslint-plugin": "^5.30.0",
+    "@typescript-eslint/parser": "^5.30.0",
+    "eslint": "^8.18.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-airbnb-typescript": "^17.0.0",
+    "eslint-config-node": "^4.1.0",
+    "eslint-config-prettier": "8.3.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "25.3.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "4.0.0",
+    "jest": "27.4.5",
+    "nodemon": "^2.0.15",
+    "prettier": "2.5.1",
+    "ts-jest": "27.1.1",
+    "ts-node": "^10.8.2",
+    "typescript": "4.5.4"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "testEnvironment": "node",
+    "testPathIgnorePatterns": [
+      "<rootDir>/dist/",
+      "<rootDir>/node_modules/"
+    ],
+    "modulePathIgnorePatterns": [
+      "<rootDir>/dist/"
+    ]
+  },
+  "scripts": {
+    "start": "node ./dist/index.js",
+    "nodemon": "nodemon --enable-source-maps ./dist/index.js --watch ./dist --watch package.json --watch ./src/views --ext js,html",
+    "compile": "tsc && mkdir -p cache",
+    "format": "prettier --config .prettierrc '{src,tests,resources}/**/*.{ts,js}' --write",
+    "verify": "npm run lint",
+    "lint": "npx eslint . --fix --ext .js,.jsx,.ts,.tsx",
+    "update": "git pull && npm install . && pm2 restart 0",
+    "watch": "tsc -w",
+    "test": "jest"
+  }
+}
