+++ old/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "esModuleInterop": true,
+    "moduleResolution": "node",
+  },
+  "include": ["/**/*.ts"],
+  "exclude": ["node_modules"]
+}
